RAG for AI Buddy/requirements.txt
langchain
langchain-ollama
langchain-core
python-dotenv
langchain-chroma
chromadb